+++ src/app/(form)/create/mentoring/page.tsx
@@ -1,123 +1,15 @@
 "use client";
-
-import { instance } from "@/service/api/instance/axiosInstance";
-
+import React from "react";
+import {useSubmit} from "@/hooks/form/useSubmit";
 import FormTitle from "../../_components/FormTitle";
 import CreateTeamForm from "../../_components/CreateTeamForm";
-
-export interface FormFields {
-  label: string;
-  name: string;
-  required?: boolean;
-  rules?: object;
-  options?: { value: string; label: string }[];
-}
-
-interface MentoringFormRow {
-  row: FormFields[];
-}
-
-type MentoringForm = FormFields | MentoringFormRow;
-
-export interface MentoringFormData {
-  name: string;
-  deadline: string;
-  startDate: string;
-  endDate: string;
-  mentoringCnt: string;
-  content: string;
-  status: string;
-  link: string;
-  role: string;
-  categories: number[];
-}
-
-const MentoringFormFields: MentoringForm[] = [
-  {
-    label: "팀 이름",
-    name: "name" as keyof MentoringFormData,
-    required: true,
-    rules: { required: "팀 이름은 필수 항목입니다." },
-  },
-  {
-    label: "모집 마감일",
-    name: "deadline" as keyof MentoringFormData,
-    required: true,
-    rules: { required: "모집 마감일은 필수 항목입니다." },
-  },
-  {
-    row: [
-      {
-        label: "멘토링 시작일",
-        name: "startDate" as keyof MentoringFormData,
-        required: true,
-        rules: { required: "멘토링 시작일은 필수 항목입니다." },
-      },
-      {
-        label: "멘토링 종료일",
-        name: "endDate" as keyof MentoringFormData,
-        required: true,
-        rules: { required: "멘토링 종료일은 필수 항목입니다." },
-      },
-    ],
-  },
-  {
-    row: [
-      {
-        label: "내 역할",
-        name: "role" as keyof MentoringFormData,
-        options: [
-          { value: "MENTOR", label: "멘토" },
-          { value: "MENTEE", label: "멘티" },
-        ],
-        required: true,
-      },
-      {
-        label: "모집인원",
-        name: "mentoringCnt" as keyof MentoringFormData,
-        required: true,
-        rules: { required: "모집인원은 필수 항목입니다." },
-      },
-    ],
-  },
-  {
-    row: [
-      {
-        label: "모집 카테고리",
-        name: "categories" as keyof MentoringFormData,
-        required: true,
-      },
-      {
-        label: "연락 방법",
-        name: "link" as keyof MentoringFormData,
-        required: true,
-        rules: { required: "연락 방법은 필수 항목입니다." },
-      },
-    ],
-  },
-  {
-    label: "팀 소개",
-    name: "content" as keyof MentoringFormData,
-    required: true,
-    rules: { required: "팀 소개는 필수 항목입니다." },
-  },
-];
-
-const defaultValues = {
-  name: "",
-  deadline: "",
-  startDate: "",
-  endDate: "",
-  mentoringCnt: "",
-  content: "",
-  link: "",
-  role: "",
-  categories: [],
-};
+import {MentoringFormFields, defaultValues} from "@/app/(form)/_data/createMentoring";
+import {MentoringFormData} from "@/app/(form)/_type/createFormData";
 
 const Page = () => {
-  const onSubmit = async (formData: MentoringFormData) => {
-    const payload = {
+  const { submit, isLoading, error } = useSubmit<MentoringFormData>({
+    endpoint: "/mentoring/teams",
+    formatPayload: (formData: MentoringFormData) => ({
       name: formData.name,
       startDate: formData.startDate,
       endDate: formData.endDate,
@@ -126,18 +18,13 @@ const Page = () => {
       link: formData.link,
       role: formData.role,
       categories: Array.isArray(formData.categories)
-        ? formData.categories.map(Number) // ✅ 모든 값을 숫자로 변환
-        : [Number(formData.categories)], // ✅ 단일 값을 숫자로 변환
-    };
-
-    console.log(payload)
-    try {
-      const response = await instance.post("/mentoring/teams", payload);
-      console.log("Response:", response.data);
-    } catch (error) {
-      console.error("Error:", error.response?.data || error.message);
-    }
-  };
+          ? formData.categories.map(Number) // ✅ 모든 값을 숫자로 변환
+          : [Number(formData.categories)], // ✅ 단일 값을 숫자로 변환
+    }),
+    onSuccess: () => {
+      alert("멘토링 팀이 성공적으로 생성되었습니다!");
+    },
+  })
 
   return (
     <>
@@ -146,11 +33,13 @@ const Page = () => {
         title="을 생성하기에 앞서 간단한 정보를 입력해주세요."
       />
       <CreateTeamForm
-        onSubmit={onSubmit}
+        onSubmit={(data) => submit(data)}
         defaultValues={defaultValues}
         formFields={MentoringFormFields}
         division={"select"}
       />
+      {error && <p className="text-red-500">{(error as Error).message}</p>}
+      {isLoading && <p className="text-blue-500">팀 생성 중...</p>}
     </>
   );
 };
